requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==3.0.7
django-nose==1.4.6
djangorestframework==3.11.0
idna==2.9
ipython-genutils==0.2.0
jedi==0.17.1
nose==1.3.7
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pinocchio==0.4.2
pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pudb==2019.2
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
urwid==2.1.0
wcwidth==0.2.4
coverage==5.1
